Fix chapter nav link
Using distinct var names for prev and next chapters. Sounds silly but I'll take it if it works.

The next chapter number computation fails, for chapter number > 1. Seems if the block to fix params of prev btn executes the plus filter fails.

Or does Liquid not know how to add 1 to anything other than 1?